+++ src/common/index.ts
@@ -1,2 +1,3 @@
+export { default as AppNavbar } from './AppNavbar'
 export { default as PageListFilters } from './PageListFilters'
 export { default as PageTitles } from './PageTitles'
