+++ src/pages/OrgNew/index.tsx
@@ -1,5 +1,5 @@
 import { Stack } from '@mui/material'
 
-export default function OrgCreate() {
+export default function OrgNew() {
   return <Stack flex={1}>Organization create page</Stack>
 }
